requirements.txt
annotated-types==0.7.0
boto3==1.35.53
botocore==1.35.53
colorama==0.4.6
greenlet==3.1.1
jmespath==1.0.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.3
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
